add create meal

+++ app/src/main/java/com/kidd/fitness/ui/insert_food/InertMealFragment.kt
@@ -40,7 +40,7 @@ class InertMealFragment : BaseFragment() {
                 Food(
                     System.currentTimeMillis().toString(),
                     edt_name.text,
-                    edt_calo.text,
+                    edt_calo.text.toInt(),
                     spinnerTimerAdapter.getItem(spinner_time.selectedItemPosition)
                     )
             )
